+++ src/gifgen/lzw.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-uint16_t lzw_compress_data(uint8_t *input, uint8_t **output, uint16_t length, uint8_t num_values);
+uint64_t lzw_compress_data(uint8_t *input, uint8_t **output, uint64_t length, uint8_t num_values);
 
 #endif
